+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/adapters/data/index.ts
@@ -1,0 +1,12 @@
+import assert from 'node:assert';
+import { DataTableMapping } from '../../core/migration-pipeline';
+import { Stack } from '@aws-sdk/client-cloudformation';
+
+export const tableMappingKey = 'DataSourceMappingOutput';
+
+export const getDataDefinition = (dataStack: Stack): DataTableMapping => {
+  const rawTableMapping = dataStack.Outputs?.find((o) => o.OutputKey === tableMappingKey)?.OutputValue;
+  assert(rawTableMapping);
+  const tableMapping = JSON.parse(rawTableMapping);
+  return tableMapping;
+};