+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/codegen-head/analytics.ts
@@ -1,0 +1,9 @@
+export type AnalyticsDimensions = Record<string, string | number | boolean>;
+
+export interface Analytics {
+  logEvent(eventName: string, dimensions?: AnalyticsDimensions): Promise<void>;
+}
+export class AppAnalytics implements Analytics {
+  constructor(private appId: string) {}
+  logEvent = async () => Promise.resolve();
+}