+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/todo_error.ts
@@ -1,0 +1,7 @@
+import ts from 'typescript';
+const factory = ts.factory;
+
+export const createTodoError = (todoMessage: string) =>
+  factory.createThrowStatement(
+    factory.createNewExpression(factory.createIdentifier('Error'), undefined, [factory.createStringLiteral(`TODO: ${todoMessage}`)]),
+  );