+++ packages/amplify-cli/src/commands/gen2-migration/shift.ts
@@ -2,6 +2,9 @@ import { AmplifyMigrationStep } from './_step';
 import { printer } from '@aws-amplify/amplify-prompts';
 
 export class AmplifyMigrationShiftStep extends AmplifyMigrationStep {
+  readonly command = 'shift';
+  readonly describe = 'NA';
+
   public async validate(): Promise<void> {
     printer.warn('Not implemented');
   }