+++ packages/amplify-console-hosting/src/utils/config-utils.js
@@ -9,6 +9,7 @@ const clientFactory = require('../utils/client-factory');
 const consolePathManager = require('../utils/path-manager');
 const buildUtils = require('./build-utils');
 const { ensureEnvParamManager } = require('@aws-amplify/amplify-environment-parameters');
+const { PutObjectCommand } = require('@aws-sdk/client-s3');
 
 function initCFNTemplate(context, templateFilePath) {
   const templateContent = context.amplify.readJsonFile(templateFilePath);
@@ -176,7 +177,7 @@ async function uploadFile(s3, filePath, key) {
     };
     const projectBucket = stateManager.getMeta().providers[constants.PROVIDER].DeploymentBucketName;
     s3Params.Bucket = projectBucket;
-    await s3.putObject(s3Params).promise();
+    await s3.send(new PutObjectCommand(s3Params));
     return projectBucket;
   }
 }