+++ packages/amplify-provider-awscloudformation/src/aws-utils/aws-cognito-client.ts
@@ -1,7 +1,7 @@
 import { $TSContext } from '@aws-amplify/amplify-cli-core';
 import { CognitoIdentityProviderClient, CognitoIdentityProviderClientConfig } from '@aws-sdk/client-cognito-identity-provider';
 import { NodeHttpHandler } from '@smithy/node-http-handler';
-import { AwsSecrets, loadConfiguration } from '../configuration-manager';
+import { loadConfiguration } from '../configuration-manager';
 import { proxyAgent } from './aws-globals';
 
 export class CognitoUserPoolClientProvider {
@@ -10,7 +10,7 @@ export class CognitoUserPoolClientProvider {
 
   static async getInstance(context: $TSContext, options = {}): Promise<CognitoUserPoolClientProvider> {
     if (!CognitoUserPoolClientProvider.instance) {
-      let cred: AwsSecrets = {};
+      let cred = {};
       try {
         cred = await loadConfiguration(context);
       } catch (e) {
@@ -22,7 +22,7 @@ export class CognitoUserPoolClientProvider {
     return CognitoUserPoolClientProvider.instance;
   }
 
-  constructor(creds: AwsSecrets, options = {}) {
+  constructor(creds, options = {}) {
     const clientConfig: CognitoIdentityProviderClientConfig = {
       ...creds,
       ...options,