+++ packages/amplify-provider-awscloudformation/src/utils/auth-types.ts
@@ -1,4 +1,5 @@
 import { $TSAny } from '@aws-amplify/amplify-cli-core';
+import { NodeHttpHandler } from '@smithy/node-http-handler';
 
 export type AuthFlow = 'admin' | 'profile' | 'accessKeys' | 'general';
 export interface AuthFlowConfig extends Partial<AwsSdkConfig> {
@@ -9,15 +10,30 @@ export interface AuthFlowConfig extends Partial<AwsSdkConfig> {
 }
 
 export interface AwsSdkConfig {
-  accessKeyId: string;
-  expiration?: Date;
+  credentials: {
+    accessKeyId: string;
+    secretAccessKey: string;
+    sessionToken?: string;
+    expiration?: Date;
+  };
   region: string;
-  secretAccessKey: string;
-  sessionToken?: string;
+  requestHandler?: NodeHttpHandler;
+  // TO DO: remove eventually, V2 style of handling proxies
   httpOptions?: {
     agent: $TSAny;
   };
 }
+export interface legacyAwsSdkConfig {
+  accessKeyId: string;
+  secretAccessKey: string;
+  sessionToken?: string;
+  expiration?: Date;
+  region: string;
+  httpOptions?: {
+    agent: $TSAny;
+  };
+}
+
 export interface AdminAuthPayload {
   accessToken: CognitoAccessToken;
   clockDrift: number;