+++ scripts/split-e2e-tests-codebuild.ts
@@ -42,6 +42,17 @@ const RUN_SOLO = [
   'src/__tests__/transformer-migrations/searchable-migration.test.ts',
   'src/__tests__/uibuilder.test.ts',
 ];
+const RUN_DUO = [
+  'src/__tests__/api_6c.test.ts',
+  'src/__tests__/auth_9.test.ts',
+  'src/__tests__/export-pull-a.test.ts',
+  'src/__tests__/export-pull-c.test.ts',
+  'src/__tests__/hosting.test.ts',
+  'src/__tests__/notifications-analytics-compatibility-in-app-2.test.ts',
+  'src/__tests__/schema-iterative-update-4.test.ts',
+  'src/__tests__/schema-searchable.test.ts',
+  'src/__tests__/studio-modelgen.test.ts',
+];
 const DISABLE_COVERAGE = [
   'src/__tests__/datastore-modelgen.test.ts',
   'src/__tests__/amplify-app.test.ts',
@@ -195,7 +206,7 @@ const splitTestsV3 = (
     const soloJobs = [];
     const osJobs = [createRandomJob(os)];
     for (let test of testSuites) {
-      const currentJob = osJobs[osJobs.length - 1];
+      let currentJob = osJobs[osJobs.length - 1];
 
       // if the current test is excluded from this OS, skip it
       if (TEST_EXCLUSIONS[os].find((excluded) => test === excluded)) {
@@ -221,6 +232,17 @@ const splitTestsV3 = (
         continue;
       }
 
+      let maxWorkers = os === 'w' ? MAX_WORKERS_WINDOWS : MAX_WORKERS;
+      if (os === 'l' && (RUN_DUO.find((duo) => test === duo) || currentJob.tests.some((duo) => RUN_DUO.includes(duo)))) {
+        maxWorkers = 2;
+        // if we had a test that requires it is in a job with only 2 tests and a job already has 2 tests, set up a new job
+        // this may mean there will occasionally be jobs that can run with 3 tests will be running with 2
+        if (currentJob.tests.length === maxWorkers) {
+          osJobs.push(createRandomJob(os));
+          currentJob = osJobs[osJobs.length - 1];
+        }
+      }
+
       // add the test
       currentJob.tests.push(test);
       if (FORCE_REGION) {
@@ -230,7 +252,6 @@ const splitTestsV3 = (
         currentJob.useParentAccount = true;
       }
 
-      const maxWorkers = os === 'w' ? MAX_WORKERS_WINDOWS : MAX_WORKERS;
       // create a new job once the current job is full;
       if (currentJob.tests.length >= maxWorkers) {
         osJobs.push(createRandomJob(os));