+++ packages/amplify-category-geo/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.24](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-geo@3.5.23...@aws-amplify/amplify-category-geo@3.5.24) (2025-10-08)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-geo
+
+
+
+
+
+## [3.5.23](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-geo@3.5.22...@aws-amplify/amplify-category-geo@3.5.23) (2025-09-17)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-geo
+
+
+
+
+
 ## [3.5.22](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-geo@3.5.21...@aws-amplify/amplify-category-geo@3.5.22) (2025-07-01)